+++ tests/NetTopologySuite.IO.GPX.Tests/GpxLongitudeTests.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+using Xunit;
+
+namespace NetTopologySuite.IO
+{
+    public sealed class GpxLongitudeTests
+    {
+        [Fact]
+        [Regression]
+        [GitHubIssue(25)]
+        public void ConstructorShouldRejectPositive180()
+        {
+            Assert.Throws<ArgumentOutOfRangeException>("val", () => new GpxLongitude(180));
+        }
+    }
+}